+++ antispam/libs/dpy.py
@@ -46,10 +46,6 @@ from antispam.dataclasses.propagate_data import PropagateData
 
 log = logging.getLogger(__name__)
 
-mark_deprecated(
-    "Using names to ignore roles and channels will be removed in 1.3.0, please use id's instead."
-)
-
 
 class DPY(Base, Lib):
     def __init__(self, handler):
@@ -169,10 +165,7 @@ class DPY(Base, Lib):
             )
 
         # Return if ignored channel
-        if (
-            message.channel.id in self.handler.options.ignored_channels
-            or message.channel.name in self.handler.options.ignored_channels
-        ):
+        if message.channel.id in self.handler.options.ignored_channels:
             log.debug("channel(id=%s) is ignored", message.channel.id)
             raise PropagateFailure(
                 data={"status": f"Ignoring this channel: {message.channel.id}"}
@@ -181,7 +174,6 @@ class DPY(Base, Lib):
         # Return if member has an ignored role
         try:
             user_roles = [role.id for role in message.author.roles]
-            user_roles.extend([role.name for role in message.author.roles])
             for item in user_roles:
                 if item in self.handler.options.ignored_roles:
                     log.debug(